+++ source/extensions/omni.flux.stage_manager.plugin.interaction.usd/omni/flux/stage_manager/plugin/interaction/usd/all_lights.py
@@ -35,7 +35,8 @@ class AllLightsInteractionPlugin(_StageManagerUSDInteractionPlugin):
         "OmniPrimsFilterPlugin",
         "SearchFilterPlugin",
     ]
-    compatible_widgets: list[str] = ["PrimTreeWidgetPlugin", "IsVisibleStateWidgetPlugin"]
+    # TODO StageManager: We have LSS plugin names in the flux ext because of this system
+    compatible_widgets: list[str] = ["PrimTreeWidgetPlugin", "IsVisibleStateWidgetPlugin", "IsCaptureStateWidgetPlugin"]
 
     async def _update_context_items_deferred(self):
         await omni.kit.app.get_app().next_update_async()