+++ source/extensions/omni.flux.stage_manager.widget/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.0]
+### Changed
+- Optimize the way the interaction UI is built
+
 ## [1.1.0]
 ### Changed
 - Deactivate interaction plugins when the tab is not visible in the widget