+++ source/extensions/omni.flux.stage_manager.window/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.0.3]
+### Changed
+- Renamed the `State` column to Actions
+
 ## [1.0.2]
 ### Added
 - Added USD listener extension as dependency for the standalone app